People: add people module.
This adds the people module which has 2 features.

The first feature is an endpoint, when following the url
`http://django-docker.dev/hello-world/<id>`, where `<id>` is an id of an
object created in the admin section, it will show the message `Hello
<name>`. `<name>` will be replaced by the name for the object in the
admin section.

The second feature is similar, but does this over the command line and
can be run with `docker exec djangodocker_web_1 python manage.py
hello-world <id>`.